Add files via upload
This project involved the first OOP implementation for solving a problem. In this case, that OOP Method was creating an Linear Feedback Shift Register class that is responsible for encrypting an image with a specific seed(which is passed to the Linear Feedback Shift register), where only that same seed can be applied on the encrypted image to decrypt it into the originally parsed image. This encryption occurs through the use of XORing each pixel of the image with seeded values. Additionally, this project was also an introduction to using BOOST cases to test if the OOP implementation was working as intended